+++ vec2.c
@@ -6,7 +6,7 @@
 /*   By: jobvan-d <jobvan-d@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/02/02 15:44:45 by jobvan-d      #+#    #+#                 */
-/*   Updated: 2022/02/08 13:36:14 by jobvan-d      ########   odam.nl         */
+/*   Updated: 2022/02/09 18:30:50 by jobvan-d      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,13 +21,13 @@ t_vec2	vec2_calc_delta(const t_vec2 a, const t_vec2 b)
 	return (delta);
 }
 
-void	vec2_divide_scalar(t_vec2 *ref, const float divisor)
+void	vec2_divide_scalar(t_vec2 *ref, const double divisor)
 {
 	ref->x /= divisor;
 	ref->y /= divisor;
 }
 
-void	vec2_multiply_scalar(t_vec2 *ref, const float multiplier)
+void	vec2_multiply_scalar(t_vec2 *ref, const double multiplier)
 {
 	ref->x *= multiplier;
 	ref->y *= multiplier;
